very rough draft
One of the big take away from the start of this project is that i never really understood just how abstract or world is. when i had the program do the computations on the income the numbers came out with out commas and with digits on running behind the decimal. I was able to use stack overflow to help with the gaps in my knowledge. There is a small amount of refactoring early on as my skin crawled when i saw how much code was repeated as i went through.
-future plans will include css and more data to display to the user.